finally

+++ src/main/java/com/homework/stream/skypro/homework/lesson25/stream/employee/Employee.java
@@ -5,11 +5,11 @@ public class Employee {
     private String surnameEmployee;
     private String lastnameEmployee;
     private int departmentEmployee;
-    private Integer salaryEmployee;
+    private double salaryEmployee;
     private final int idEmployee;
     private static int idCounter = 0; // Счечик для выдачи id сотруднику
 
-    public Employee(String surnameEmployee, String nameEmployee, String lastnameEmployee, int departmentEmployee, Integer salaryEmployee) {
+    public Employee(String surnameEmployee, String nameEmployee, String lastnameEmployee, int departmentEmployee, double salaryEmployee) {
         this.surnameEmployee = surnameEmployee;
         this.nameEmployee = nameEmployee;
         this.lastnameEmployee = lastnameEmployee;
@@ -57,11 +57,11 @@ public class Employee {
         this.departmentEmployee = departmentEmployee;
     }
 
-    public Integer getSalaryEmployee() {
+    public double getSalaryEmployee() {
         return salaryEmployee;
     }
 
-    public void setSalaryEmployee(Integer salaryEmployee) {
+    public void setSalaryEmployee(double salaryEmployee) {
         this.salaryEmployee = salaryEmployee;
     }
